some changes to the way the plugin behaves when fuse is not available.

+++ rplugin/python3/nvimfs/__init__.py
@@ -10,8 +10,13 @@ from nvimfs.fs import __file__ as fs_script
 @neovim.plugin
 class NvimFSHandler(object):
     def __init__(self, vim):
+        self.fuse_available = True
         self.vim = vim
         self.config = None
+        try:
+            import nvimfs.fuse
+        except OSError:
+            self.fuse_available = False
 
     def __lshift__(self, args):
         """
@@ -27,11 +32,14 @@ class NvimFSHandler(object):
         """
         Setup the plugin on start.
         """
-        self.__current_client = None
-        self.config = NeovimFSDict(self.vim)
-        self.config['id'] = self.vim.channel_id
-        self.config['mountpoint'] = join(self.vim.eval('&rtp').split(',')[0], 'neovimfs')
-        self.spawn_neovimfs()
+        if self.fuse_available:
+            self.__current_client = None
+            self.config = NeovimFSDict(self.vim)
+            self.config['id'] = self.vim.channel_id
+            self.config['mountpoint'] = join(self.vim.eval('&rtp').split(',')[0], 'neovimfs')
+            self.spawn_neovimfs()
+        else:
+            self << 'echoerr "nvimfs: couldn\'t load fuse"'
 
     def spawn_neovimfs(self):
         """
@@ -52,15 +60,16 @@ class NvimFSHandler(object):
 
     @property
     def current_client(self):
-        if not self.__current_client:
-            client_names = glob(join(self.config['mountpoint'], 'clients', '*', 'name'))
-            for n in client_names:
-                with open(n, 'r') as cn:
-                    name = cn.read()
-                    if name == environ['NVIM_LISTEN_ADDRESS']:
-                        self.__current_client = dirname(n)
-                        break
-        return self.__current_client
+        if self.fuse_available:
+            if not self.__current_client:
+                client_names = glob(join(self.config['mountpoint'], 'clients', '*', 'name'))
+                for n in client_names:
+                    with open(n, 'r') as cn:
+                        name = cn.read()
+                        if name == environ['NVIM_LISTEN_ADDRESS']:
+                            self.__current_client = dirname(n)
+                            break
+            return self.__current_client
 
     @neovim.autocmd('VimLeavePre', pattern='*')
     @neovim.shutdown_hook
@@ -68,12 +77,12 @@ class NvimFSHandler(object):
         """
         Remove the current client tree from the Oxberry filesystem when vim exits.
         """
-        if self.current_client:
+        if self.fuse_available and self.current_client:
             rmtree(self.current_client)
 
     @neovim.autocmd('BufAdd', eval='fnamemodify(expand("<afile>"), ":p")')
     def add_buffer(self, filename):
-        if self.current_client and isfile(filename):
+        if self.fuse_available and self.current_client and isfile(filename):
             buffer_nr = str(self.vim.eval('bufnr("$")'))
             path = join(self.current_client, "buffers", buffer_nr)
             mkdir(path)
@@ -83,17 +92,18 @@ class NvimFSHandler(object):
 
     @neovim.autocmd('BufDelete', eval='fnamemodify(expand("<afile>"), ":p")')
     def remove_buffer(self, filename):
-        buf_tree_to_delete = None
-        if self.current_client:
-            buffer_names = glob(join(self.current_client, 'buffers', '*', 'name'))
-            for bn in buffer_names:
-                with open(bn, 'r') as bufname:
-                    name = bufname.read()
-                    if name == filename:
-                        buf_tree_to_delete = dirname(bn)
-                        break
-            if buf_tree_to_delete:
-                rmtree(buf_tree_to_delete)
+        if self.fuse_available:
+            buf_tree_to_delete = None
+            if self.current_client:
+                buffer_names = glob(join(self.current_client, 'buffers', '*', 'name'))
+                for bn in buffer_names:
+                    with open(bn, 'r') as bufname:
+                        name = bufname.read()
+                        if name == filename:
+                            buf_tree_to_delete = dirname(bn)
+                            break
+                if buf_tree_to_delete:
+                    rmtree(buf_tree_to_delete)
 
     @neovim.autocmd('BufEnter', pattern='*/clients/*/eval,*/clients/*/cmd')
     def handle_eval_files(self):
